Fix case where Traefik can read empty plugin configuration file.
Fixed a case where Traefik can read an empty or missing file when reading the Service Fabric configuration file generated by the plugin, with some unlucky timing.
Traefik reads the configuration file every few seconds, and this can randomly lead to Traefik thinking that there are no Service Fabric resources to route to.
This will lead to all requests to that Traefik instance failing for a few seconds until the next time Traefik reads the configuration.

Changed the configuration file to be written to a temporary file first, and only then replacing the configuration file that Traefik reads.
This replacing should be an atomic operation, preventing Traefik from having a time when it can read an empty or missing configuration.
I've tested that in practice the empty configuration file case no longer happens with this fix.